es: dynamically determine doc type
Check the index if a doctype is already in use. Default to
"doc" if not.

Required for ES6 as an index can only have one doctype. And as
Logstash switched from "log" to "doc", EveBox needs to as well
to keep compatibility.

The issue is the index on the day of the upgrade, as it will still
be using the old doc type until the next days index is created.