event: if host is an object, use host.name
Suricata will create events with the "host" field being a string
of the sensor name. Filebeat will override this with a "host"
object where "host.name" is the hostname of the machine
generating the alert. If this appears to be the case, use host.name
as the sensor name.

Github issue:
https://github.com/jasonish/evebox/issues/100