elastic: lower bucket sizes
Lower the size of the aggregations buckets for the inbox query
to go over Elastic limits which seem to be more strictly enforced
in newer versions.

This may mean that not all alerts are provided in the inbox view,
until some are archived to lower the total number of alerts that
need to be aggregated, as the "terms" query is for top items,
not all.

TODO: Research the composite index which should get around this
  issue.

GitHub issue:
https://github.com/jasonish/evebox/issues/180