auto-archive: Archive events if specific metadata occurs
If an event contains the alert metadata {"evebox-action": "archive"},
the server will automatically mark the event as archived.

This works in conjunction with the new (experimental) Suricata-Update
feature to add metadata using its own rule matching.

GitHub issue: https://github.com/jasonish/evebox/issues/52

This is only done at the server for now and for Elasticsearch. It
requires that events be forwarded through the server and will not work
with Logstash, Filebeat, etc.