logging: update tokio_tracing
Removes one dependency on chrono. This uses the Offset time formatter,
which means it will log in local time, until a daylight savings event
happens, which will be ignored and the old time offset will continue to
be used.

This is due to limitations in Rust's ability to get the current local
time.