server: @before and @after querystring fields
Treat key/value pairs in the query string with the keying being
@before or @after as special time range variables.

If @before is used the time will be as as a less than or equal to
filter. Likewise, the value after @after will be used as a greater
than or equal.

The "equal to" behaviour was chosen for simplicity to avoid having to
add another set of keywords for that behaviour.