server/querystring: convert to a stage parser
This is mainly to simplify error handling with Nom.  Due a first pass
with Nom, and Nom errors that tokenizes the query string into simple
types like strings and key value pairs.

Then a second pass that will upgrade special keywords to their
specific element type such as @ip, and in the future @before and @after.