elastic/histogram: server side interval selection
When an early time is not provided and the database is used for the
early time, overide the interval with something that makes more sense.

For example, if the client selects "1d" intervals, but there are only
a few hours in the database.

TODO: Make the interval in the API optional, allowing the client to
force the interval, otherwise letting the server side choose.