elastic: handle eve dns v3
Use runtime mappings to expose a new field, "dns_query_rrname".  For
v2 logs this maps to "dns.rrname", and for v3 logs it maps to
"dns.queries.rrname".

Elastic appears to flatten the queries array which make it simple.

Runtime mappings are only applied to event query and agg queries, but
need to make it apply to all queries to be complete.

Issue: https://github.com/jasonish/evebox/issues/317