elastic: more permissive error handlings
First, we were getting a panic trying to access an object key that
didn't exist for some errors, like older OpenSearch.

Second, something was up with the usage of anyhow and thiserror
causing a panic while display the error.

Instead, just check if the response is an error and log the whole
thing.