webapp: Kibana inspired +/- filters
Add buttons to add filters for src_ip, dest_ip and
signature_id. Filters can be filter for, or filter out and are added
to a for of filters instead of the query string.

Still an issue with SolidJS "reactivity" where the result set isn't in
sync with the filters, but it might be related to hot-reload during
development.