general: more response alerts and reports with SSE
SSE (server side events) allows us to stream partial results to the
client. This can really help with responsivement on large amounts of
data in SQLite.

For Elastic, we fake it by just sending a single SSE response as Elastic
is responsive enough on large datasets.

Plus a bunch of misc. changes across the board.