sqlite: bring back rusqlite for testing
SQLx, while great to use bolts-on async to SQLite which is not async
itself. This adds overhead that can be noticed in batch operations
such as bulk inserting, or perhaps row-by-row retrieval.

Currently its opt-in, but will make more sense to use when I convert
event submission to an actor system for better batching.